fix and ref: improved the buffers
